mejora en pantalla de registro, y mejora en pantalla login
