+++ package.json
@@ -18,6 +18,7 @@
     "expo-status-bar": "~1.6.0",
     "react": "18.2.0",
     "react-native": "0.72.6",
+    "react-native-keyboard-aware-scroll-view": "^0.9.5",
     "react-native-navigation": "^7.37.2",
     "react-native-screens": "^2.10.1"
   },
